Extract handler.
