+++ src/commands/junit-upload.ts
@@ -1,12 +1,7 @@
-import { readFileSync } from 'fs'
 import { Arguments, Argv, CommandModule } from 'yargs'
-import { JUnitResultType, JUnitTestCase, parseJUnitXml } from '../utils/junit/junitXmlParser'
-import { Api, createApi } from '../api'
-import chalk from 'chalk'
-import { parseApiToken, parseUrl, twirlLoader } from '../utils/misc'
-import { ResultStatus, RunTCase } from '../api/schemas'
+import { parseUrl } from '../utils/misc'
 import { API_TOKEN } from '../config/env'
-import { dirname } from 'path'
+import { JUnitCommandHandler } from '../utils/junit/JUnitCommandHandler'
 
 export interface JUnitArgs {
 	subdomain: string
@@ -101,146 +96,7 @@ export class JUnitUploadCommandModule implements CommandModule<unknown, JUnitArg
 	}
 
 	handler = async (args: Arguments<JUnitArgs>) => {
-		const apiToken = parseApiToken(args)
-		const url = parseUrl(args)
-		const api = createApi(url, apiToken)
-		const file = readFileSync(args.file).toString()
-		const { testcases: junitResults } = await parseJUnitXml(file, dirname(args.file))
-		const tcases = await api.runs.getRunTCases(args.project, args.run).catch(printErrorThenExit)
-		const { results, missing } = mapTestCaseResults(junitResults, tcases)
-
-		validateAndPrintMissingTestCases(args, missing)
-		validateAndPrintMissingAttachments(args, results)
-
-		await uploadTestCases(args, results, api)
-
-		console.log(`Uploaded ${results.length} test cases`)
-	}
-}
-
-const validateAndPrintMissingAttachments = (
-	args: Arguments<JUnitArgs>,
-	results: TCaseWithResult[]
-) => {
-	if (args.attachments) {
-		let hasAttachmentErrors = false
-		results.forEach(({ result }) => {
-			result.attachments.forEach((attachment) => {
-				if (attachment.error) {
-					printError(attachment.error)
-					hasAttachmentErrors = true
-				}
-			})
-		})
-		if (hasAttachmentErrors && !args.force) {
-			process.exit(1)
-		}
-	}
-}
-
-const validateAndPrintMissingTestCases = (args: Arguments<JUnitArgs>, missing: JUnitTestCase[]) => {
-	missing.forEach((item) => {
-		const folderMessage = item.folder ? ` "${item.folder}" ->` : ''
-		const header = args.force ? chalk.yellow('Warning:') : chalk.red('Error:')
-		console.error(
-			`${header}${chalk.blue(`${folderMessage} "${item.name}"`)} does not match any test cases`
-		)
-	})
-	if (missing.length && !args.force) {
-		process.exit(1)
-	}
-}
-
-const uploadTestCases = async (
-	args: Arguments<JUnitArgs>,
-	results: TCaseWithResult[],
-	api: Api
-) => {
-	const loader = twirlLoader()
-	loader.start()
-	try {
-		for (let i = 0; i < results.length; i++) {
-			const { tcase, result } = results[i]
-			let comment = result.message
-			loader.setText(`Uploading test case ${i + 1} of ${results.length}`)
-			if (args.attachments) {
-				const attachmentUrls: Array<{ name: string; url: string }> = []
-				for (const attachment of result.attachments) {
-					if (attachment.buffer) {
-						const { url } = await api.file.uploadFile(
-							new File([attachment.buffer], attachment.filename)
-						)
-						attachmentUrls.push({ url, name: attachment.filename })
-					}
-				}
-				comment += `\n<p>Attachments:</p>\n${makeListHtml(attachmentUrls)}`
-			}
-
-			await api.runs.createResultStatus(args.project, args.run, tcase.id, {
-				status: getResult(result.type),
-				comment,
-			})
-		}
-		loader.stop()
-	} catch (e) {
-		printErrorThenExit(e)
-	}
-}
-
-const makeListHtml = (list: { name: string; url: string }[]) => {
-	return `<ul>
-${list.map((item) => `<li><a href="${item.url}">${item.name}</a></li>`).join('\n')}
-</ul>`
-}
-
-const getResult = (result: JUnitResultType): ResultStatus => {
-	switch (result) {
-		case 'error':
-			return 'blocked'
-		case 'failure':
-			return 'failed'
-		case 'skipped':
-			return 'skipped'
-		case 'success':
-			return 'passed'
-	}
-}
-
-interface TCaseWithResult {
-	tcase: RunTCase
-	result: JUnitTestCase
-}
-
-const mapTestCaseResults = (junitTCases: JUnitTestCase[], testcases: RunTCase[]) => {
-	const results: TCaseWithResult[] = []
-	const missing: JUnitTestCase[] = []
-
-	junitTCases.forEach((result) => {
-		const tcase = testcases.find(
-			(tcase) => tcase.title === result.name && result.folder === tcase.folder.title
-		)
-		if (tcase) {
-			results.push({
-				result,
-				tcase,
-			})
-			return
-		}
-		missing.push(result)
-	})
-
-	return { results, missing }
-}
-
-const printErrorThenExit = (e: unknown): never => {
-	printError(e)
-	process.exit(1)
-}
-
-const printError = (e: unknown) => {
-	if (e instanceof Error) {
-		console.error(e)
-	} else {
-		console.error(e)
+		const handler = new JUnitCommandHandler(args)
+		await handler.handle()
 	}
 }
